calculator, CarbCalcDto.java minor refactor calculator test add

+++ src/main/java/com/raspberrypi/fermzilla/rPiMSMonitor/dataProcessor/CarbCalcDto.java
@@ -1,7 +1,6 @@
 package com.raspberrypi.fermzilla.rPiMSMonitor.dataProcessor;
-import java.math.BigDecimal;
 
-public record CarbCalcDto(double desiredCarbonation, int beerTemperature, BigDecimal requiredPressureInBars) {
+public record CarbCalcDto(double desiredCarbonation, int beerTemperature, double requiredPressureInBars) {
 
     @Override
     public String toString() {
